Small changes to UI for clarity/ease
Appended some text/titles for each of the boxes to make clearer the 1st kinetic input and later kinetic inputs.

Also, relevant things grey out when 'Use period before pulse for first background' is checked, and it starts unchecked by default.

Finally, the delimiter sets itself to a comma by default, as it seems to be what people use when converting from `.sif` to `.asc`.